refactor(shared): extract authenticated request context
