+++ src/main/java/com/renewsim/backend/shared/security/AuthenticatedRequestContext.java
@@ -1,0 +1,6 @@
+package com.renewsim.backend.shared.security;
+
+public record AuthenticatedRequestContext(
+        String username,
+        boolean isAdmin) {
+}
