--- src/main/java/com/renewsim/backend/simulation_service/shared/web/SimulationRequestContext.java
@@ -1,6 +1,0 @@
-package com.renewsim.backend.simulation_service.shared.web;
-
-public record SimulationRequestContext(
-                String username,
-                boolean isAdmin) {
-}
